clap: Initial commit
This is an attempt at a small, lightweight, cross-platform 3D game engine
that I'm going to write in C, because learning Rust will take too long to
be practical.

Despite being plain C, it's not too procedural and will gradually become
even less so. The procedural bits are a side effect of trying to get
unfamiliar things to work and getting confused.

Dependencies are minimal. Native build requires GLFW and GLEW. Emscripten
build doesn't require anything. Both do the same thing. Everything is
pretty much all over the place, but that's because early days.

Signed-off-by: Alexander Shishkin <alex@ash.works>