Massive 'everything' commit
Stuff that I've been hacking on for the last half a year or so. It's all
over the place. A lot of things reworked. A lot of new stuf added:

 * ui widgets,
 * ui animations,
 * basic networking (native sockets and websockets),
 * basic physics via ODE.

I will probably go ahead and squash everything into one 'initial commit'
commit to set the ground zero. Things need to be separated enough so that
hacking on one aspect doesn't have to span half the source files.

Signed-off-by: Alexander Shishkin <alex@ash.works>