input-joystick: Factored out gamepad code
Some kind of a common ground between GLFW's joystick API and Emscripten's
gamepad API. Hacky in a few places: controller mapping in GLFW code is
completely arbitrary (taken from DualShock controller on linux). Mapping
itself is done via macros because I want to keep the native build working
the same way as before this change.

The wasm build now responds to DualShock the same way as the native build,
plus, PS3 controller responds the same way, unlike in the native build.

This is rushed, because LDjam#48 starts tomorrow and the wasm build is an
obvious priority.

One thing to do would be to upgrade from GLFT's joystick API to their
gamepad API, which has a build-in table of game controller mappings.

Signed-off-by: Alexander Shishkin <alex@ash.works>