object: Add static objects
Static objects are not allocated or freed, reference manipulation on them
is illegal. Instead, they are initialized by ref_embed() inside a larger
object that's either static or allocated. In the latter case it's expected
to have its own refcounting.

Don't include them in the stats and complain loudly if the
assumption is broken.

While at it, move all static type information to refclass and clean up the
struct ref.

Signed-off-by: Alexander Shishkin <alex@ash.works>