Add support for opaque types
In the quest to (not) reinvent C++, add support for opaque types such that
only the compilation unit that defines IMPLEMENTOR can access type's data
structure. Not sure yet if this is valuable, but it's at least relatively
cheap.

The assumption that there will be one such type per header file.

The implementor then will have to set IMPLEMENTOR around the include of
that header file, while everybody else includes it the regular way. This
may have implications in debuggers, but it's also trivial to disable the
opacity if need be.

Signed-off-by: Alexander Shishkin <alex@ash.works>