render/gl: Begin the solemn work of abstracting away OpenGL
The first order of business is to wrap all GL calls in something that
checks for errors and lets us know where problems are. While at it,
we might as well factor out common code sequences and group all relevant
bits in our own data structures.

We start with struct texture / texture_t, which replaces GUints in the
rest of the code; texture_t becomes the basic texture unit instead.

When a different renderer comes along, it will only have to deal with
replacing GL bits inside texture_t with its own, instead of trawling
through the entire code base.

Signed-off-by: Alexander Shishkin <alex@ash.works>