mesh: A new type suitable for vertex mesh manipulation
There are places in the codebase where a model's vertex data is passed
around as disjointed vertex/index arrays, which is not ideal. At the same
time, the impending use of meshoptimizer suggests that there has to be a
generalized way of dealing with vertex data.

Introduce a new mesh type that encapsulates all the different vertex /
index / texture coordinate / etc arrays and add a basic "optimize"
operation that performs a number of meshoptimizer's tricks on its data.

Start by converting the gltf loader to use struct mesh.

Signed-off-by: Alexander Shishkin <alex@ash.works>