Make proper use of gltf animation data
gltf is structured in a way that doesn't really have 'poses' as in global
snapshots of all the transforms at a given point in time. Instead, the
animation data is split between channels, each one driving a path (T/R/S)
of a given joint, each one with its own timeline.

Signed-off-by: Alexander Shishkin <alex@ash.works>