input: Rework keyboard input
First, upgrade input message fields from bits to bytes; the message gets
bigger, but it's ok, they are stack allocated anyway.

Second, move the entire mapping logic to a table where it's easier to
manage and potentially configure in the future.

While at it, clearly differentiate between presses and holds.

Signed-off-by: Alexander Shishkin <alex@ash.works>