character: Move all motion out of input parsing
There are multiple problem with mixing motion with input parsing, most
of which boil down to the fact that we receive multiple input messages
per frame. The cure is to first collect all the input messages and then
act on them.

This fixes the jittering camera when controlled from the keyboard.

Signed-off-by: Alexander Shishkin <alex@ash.works>