mesh: Handle 32-bit index width slightly better
Because GLES/WebGL can only deal with 16-bit index arrays, we're still
limited to 64k vertices, but if libmeshoptimizer manages to reduce the
incoming vertices to below 64k, we can still work with that.

To that end, don't convert the index array to 32-bit indices for the
meshoptimizer, if the incoming index array is already 32-bit width.

The reverse conversion still has to happen, and if any index is above
USHRT_MAX, abort.

TODO: replace the abort with proper error propagation, as the rest of
the engine can still function if a model fails to load.

Signed-off-by: Alexander Shishkin <alex@ash.works>