gltf: Don't crash on unsupported texture formats
Only "image/png" is supported at the moment, anything else leads to a
segfault on dereferencing an empty array of images. Add such images as
"unsupported" and try to continue loading the model, in the hopes that
these images are not used anywhere. If they are, communicate an error up
the call chain, abort the gltf model load and move on.

Signed-off-by: Alexander Shishkin <alex@ash.works>