model: Drop the passed reference to model on failure
When calling model3dtx_new_*() with ref_pass(model) and it fails, the
caller still has to clean up the last reference of model, even though
(and it's also a hint) the model at that point would be NULL. Currently,
we're employing a hack around that.

Instead of this, drop the passed reference to model in model3dtx_new_*()
functions directly.

Signed-off-by: Alexander Shishkin <alex@ash.works>