ldjam56: Move scene initialization below clap_init()
There is a circular dependency between clap_init() and scene_init(),
which at the moment is dealt with by the latter coming before the
former, which in turn results in all sorts of inconveniences.

Move scene_init() after clap_init(). While at it, make exit_timeout
out of struct scene and local to the main compilation unit.

Signed-off-by: Alexander Shishkin <alex@ash.works>