character: Fix motion vector calculation
The motion vector calculation is broken in 2 places: first, the new
coordinate system's unit vectors are scaled by the motion vector
components twice, which resulted in an unintentionally convoluted unit
vector calculation and as a byproduct resulted in slower diagonal
motion. Fix this and make the end result actually make sense.

Signed-off-by: Alexander Shishkin <alex@ash.works>