Replace preprocess_shaders.c with a shell script
Turns out, if we use #version 300 es profile, we don't need to make so
many changes to the shaders at all. The ES is still the lowest common
denominator, so shaders need to take it into account, but the same
shaders will work just fine with core profile. The only change we need
to make is the version header and add float precision declaration for
the ES. This is easier done with sed from a shell script, which is also
easier to maintain. The downside is that whoever ports this to windows
will probably have to rewrite it in python.

Signed-off-by: Alexander Shishkin <alex@ash.works>