pipeline, fbo: Support multiple texture attachments
For bloom effect, emission output needs to be rendered to a separate
texture in the framebuffer that can then be post-processed separately
from the main one and then combined back together.

This requires support for multiple color attachments in the framebuffer
object and multiple textures in render pipeline passes.

While at it, also implement ping-pong render passes (for multiple stages
of blur, in this instance, for better effect).

This also lays the foundation for deferred rendering.

To make use of this functionality, add bloom support to the main
rendering pipeline in demo/ldjam56.

Signed-off-by: Alexander Shishkin <alex@ash.works>