Use a dedicated texture as a sobel input
Problem 1: lighting calculations mess up with edge detection, which
hops around while characters move.
Problem 2: outlining edges on the character is not ideal, it's better to
just outline the character and its texture elements.
Problem 3: faces with close enough color aren't picked up by the edge
detection, leaving, for example dark foliage or basalt pillars without
outlines.

Solve these by generating a dedicated texture for the edge detection
pass instead of applying it to a fully rendered and lighted scene. In
it, use a unique entity hash as a seed for coloring the characters.
Instead of textured faces, use their normal vectors as colors, which
makes adjacent faces of the same color, but at an angle to each other
appear as different colors to the edge detection pass, solving problem
3. This is still not ideal as parallel faces with the same normal not
have an edge between them, but outlines geometries much better.

Signed-off-by: Alexander Shishkin <alex@ash.works>