Add ImGui via cimgui bindings
ImGui is too good not to make use of it. Instead using cimgui/imgui
build infrastructure, grab the compilation units that we actually
need and build them directly into the engine library. This requires
a bit of hackery because of C++ interop and a slight on-the-fly
modification to backend headers.

Signed-off-by: Alexander Shishkin <alex@ash.works>