ui-imgui: Add basic initialization/deinitialization and renderning
For now, draw ImGui's demo window on top of the scene to see that
it works. Hook up the initialization/deinitialization to the glfw
part of the code. The same is done for emscripten, except don't
use the glfw emulation, as it's just one more intermediate layer.
Instead, skip the platform backend for now. All that's really needed
is input, and mostly mouse input. When enough of that is done, it
can be made into a proper backend for ImGui.

Signed-off-by: Alexander Shishkin <alex@ash.works>