ldjam56: Move ui initialization after pipelines
If ui initialization fails, we can still have a somewhat functional
render loop in the browser if we have the pipelines, because the main
loop will soldier on even if main() fails. Letting the pipelines
initialize before the ui potentially fails allows the render_frame()
to not crash, so do that.

Signed-off-by: Alexander Shishkin <alex@ash.works>