ldjam56: Use glslc and spirv-cross to generate shaders
The need for generating different versions of shaders for different
render backends existed from day one. The homespun scripts can only go
so far. On the other hand, there already exists a toolchain for
translating shaders from GLSL to practically anything, using SPIR-V as
an intermediate representation, including GLSL to GLSL ES and GLSL to
GLSL, which is what is needed at the moment, but in the future can
produce shaders for other rendering backends.

Use this toolchain to generate shaders for runtime.

Signed-off-by: Alexander Shishkin <alex@ash.works>