shader: Get rid of parsing shader sources for variable names
It's not clear why this exists in the first place, but it's very
simplistic and buggy and is entirely unnecessary, since all the
attributes and uniforms that need to be used in the code are known and
are, basically, all of them. The end result is more calls to
glGetAttribLocation()/glGetUniformLocation(), as all known attributes
and uniforms have to be probed for, but that still only happens at
startup time, so not a big problem.

The bugginess is the inability to handle precision specifiers between
a type name and a uniform or in/out name. This means that shaders
generated by spirv-cross will not work with this scheme. The more
complex the shaders become, the more kludges this code would require.

Get rid of the entire shader parsing in favor of direct introspection GL
calls.

Signed-off-by: Alexander Shishkin <alex@ash.works>