shader: Rework handling of attribute arrays, uniforms and textures
There are lots of repeating patterns of GL calls in model.c that create,
upload, set up, configure (and the reverse) attribute arrays, uniforms
and textures. All this code pokes directly at the locations stored in
the shader_prog structure, under names that often don't match the
corresponding names in the shaders. Instead, create a mapping between
attributes and uniforms to constants and provide helpers for all the
open coded operations on them. This moves a lot of GL calls into
shader.c, where validation and error handling is done in one place and
makes the code instantly more readable. This will also make it easier to
introduce uniform buffer objects in the future and decouple the model
and shader code from the GL render backend when the time comes.

Signed-off-by: Alexander Shishkin <alex@ash.works>