model: Streamline USE_NORMALS setting
It would appear that models_render() only sets USE_NORMALS (to either
true or false) if there are normals, that is, txm->normals is not NULL.
This would mean that USE_NORMALS would be true for models that don't
have normals, but follow a model that does. In reality, txm->normals is
never NULL, so it works, but the NULL check is redundant.

Signed-off-by: Alexander Shishkin <alex@ash.works>