ui: Use a different quad for UI elements
UI coordinate system is upside down, unlike the NDC and simply reverting
it is not enough, because the triangle faces point away from the screen,
which necessitates disabled face culling. To avoid all that, use a
reverse quad and enable back the face culling. While at it, also disable
depth testing for the UI elements.

Signed-off-by: Alexander Shishkin <alex@ash.works>