util: Provide a rough approximation of the pieces missing from wlibc
windows-libc doesn't do srand48()/drand48()/lrand48() that and the terrain
generator uses them quite extensively. These are made just so that stuff
would compile and should be implemented properly, preferrably in wlibc.

PATH_MAX is also missing. I have no idea what it should be on windows,
so just go with what I like. Somebody should look into this. And it also
should probably be in wlibc instead.

M_PI and friends should also be somewhere in the general vicinity of
math.h.

Signed-off-by: Alexander Shishkin <alex@ash.works>