model, pipeline: Add an override shader to the renderer
When rendering into a depth map, regular model shaders are unnecessary,
as the process doesn't produce any color outputs, so all that's really
needed is correct vertex transforms (including joint transforms) and an
empty fragment shader. Add an option to pass such a shader to override
models' shaders for depth map rendering. Add an extra parameter to the
pipeline setup to pass this shader to the rendering function.

Signed-off-by: Alexander Shishkin <alex@ash.works>