model, pipeline: Use a depth map texture to create shadows
Depth map texture contains Z component for every fragment rendered from
the light source's point of view. For every fragment rendered from the
camera's point of view, check its Z component in light space against the
depth map to determine if it's in the shadow.

Only light source 0 does this at the moment.

Allow a render pass to include a depth map as one of the sources.

Signed-off-by: Alexander Shishkin <alex@ash.works>