pipeline: Make UI dropdown select individual sources of each pass
Some render passes have multiple sources, so it is useful to be able to
see into each of them instead of just the first one. Make them all
selectable. This is suboptimal, as we have to iterate all passes and
sources twice each frame, it would be better to render the list at
pipeline_add_pass() time, but so far this code doesn't show up in the
perf report, so I'll worry about it later.

Signed-off-by: Alexander Shishkin <alex@ash.works>