render-gl: Make FBO depth texture attachment clamp to white border
This should be configurable via FBO parameters, in theory, but for now
the only wrapping mode that makes sense for depth textures is
clamp-to-border (or clamp-to-edge on GL ES), with the border set to
solid white. That is, the area outside of the shadow map should not be
in shadow.

Signed-off-by: Alexander Shishkin <alex@ash.works>