view: Split view into "subviews"
For cascade shadow mapping, the camera frustum needs to be split into
cascades, which become camera's subviews, each corresponding to a light
frustum that encloses that cascade, these will be the light subviews.
The total of camera frustum becomes the "main" subview.

Perform the split and convert the code to use the main subview of a
view, so that functionally everything stays the same.

Signed-off-by: Alexander Shishkin <alex@ash.works>