shaders/model: Sample shadow cascade corresponding to the distance
The cascade to sample depends on the distance of the fragment being
rendered from the camera. This is obtained by translating the fragment's
position into the camera space and comparing the result's Z component
against cascade_distances[] array. The lowest matching element is the
cascade index that needs to be sampled.

Use the light space matrix (shadow_mvp[]) of the corresponding cascade
to translate the fragment into light space and compare the resulting Z
component against the corresponding point on the depth map.

Use a switch statement to select the shadow map, as in the GL ES case
they're all separate sampler2D textures. For the GL case, just pass the
coordinates and cascade layer to the sampling function.

Signed-off-by: Alexander Shishkin <alex@ash.works>