render-gl: Replace fbo_new()/fbo_new_ms() with a new fbo_new()
The fbo constructor is growing more parameters, a lot of them are
optional, so make fbo_new() take a initializer structure instead.
The structure has width and height at the top, so the old fbo_new()
users don't have to be converted as fbo_new(width, height) would mean
the same thing as before.

Signed-off-by: Alexander Shishkin <alex@ash.works>