ui-debug: Factor out ImGui debug window operations
There are 3 booleans that each debug_module user has to deal with:
whether to display the window at all, whether it should be closed and
whether its content is folded.

Add ui_igBegin(), ui_igBegin_name() and ui_igEnd() to take care of all
that so the callers don't have to. Change the existing callers to this
new API.

Signed-off-by: Alexander Shishkin <alex@ash.works>