ui: Fix (most of) the compiler warnings
It would be useful for the development process to enable -Wall -Werror
for the engine core compilation units eventually, as the compiler is
more often right than not and all help is good. Also, ui.c is long due
for an overhaul. For now, get rid of the most of (unused something,
set-but-not-used something etc) warnings.

Signed-off-by: Alexander Shishkin <alex@ash.works>