physics: Remove phys_body_is_grounded()
phys_body_is_grounded() is only used in character_is_grounded(), which
has been unused for quite some time. Currently, character::ragdoll is
set to inverse of phys_body_ground_collide(), which does the same thing
plus it corrects the height of the character. This whole flow needs
rework and ragdoll is obviously a misnomer. But first, remove the unused
functions.

Signed-off-by: Alexander Shishkin <alex@ash.works>