ui: Fix (false positive) potentially uninitialized warning
Admittedly, not an obvious false positive, because the variables in
question are set and used under conditions that mean the same thing. And
again, only GCC seems to have a problem with it. Unlike the json.c case,
it's not a good idea to disable any warnings on ui.c, because it's a
shifty customer and it's good to catch any warnings there as they
appear. So, explicitly initialize these variables to appease GCC.

Signed-off-by: Alexander Shishkin <alex@ash.works>