cmake: Silence a false positive GCC warning in json.c
GCC detects a false positive maybe-used-uninitialized in parse_string()
in json.c. One way of dealing with it is to explicitly initialize the
string buffer in both (out and !out) cases, but because there's no real
problem there, it only matters to GCC and it's preferrable not to touch
json.c needlessly. The other option is to disable -Wmaybe-uninitialized
for json.c. Do the latter.

Signed-off-by: Alexander Shishkin <alex@ash.works>