render/gl: Replace white/black/transparent pixels with functions
Exporting texture_t symbols from render.h trips up -flto optimizations,
because it's different for render-gl.c and the rest of the code, making
it effectively aliasing. In other compilation modules, the embedded
texture_t usage is consistent, but not across the render-gl.c boundary.

Replace the exported symbols with functions that return pointers to
texture_t of the corresponding pixels, which remain local to render-gl.

Fix the pixel textures' users to use these functions instead.

Signed-off-by: Alexander Shishkin <alex@ash.works>