physics: Fix character springboarding off of uneven terrain
In phys_body_ground_collide(), cast a longer ray down and unless the
body is airborne, stick it to the ground. While at it, make the ray
collision be the closest to its origin, because that's the useful one.
Also, remove height correction in phys_step() using
phys_body_ground_collide(), which is effectively undoing the collision
correction there.

Signed-off-by: Alexander Shishkin <alex@ash.works>