physics: Get rid of the global physics state
Make physics state part of the clap_context, add an accessor to obtain
it from the context, otherwise obtain it from the phys_body. Change the
callers to adjust.

The only global physics state is the phys_bodies list, which is
currently not used for anything, but now is used as a hack to obtain the
physics context for phys_ray_cast() for entities that don't have
phys_body, but still want to cast rays.

Signed-off-by: Alexander Shishkin <alex@ash.works>