physics: Split phys_ray_cast()
Currently, phys_ray_cast() only works for entities that have a
phys_body, which is a problem for terrain clamping of entities that
don't. For this reason, a hack is used to obtain the physics state.

Get rid of the hack and add a version of phys_ray_cast() that takes a
physics state as a parameter.

Signed-off-by: Alexander Shishkin <alex@ash.works>