gltf: Implement GLB loading
According to Khronos Group, GLTF binary (GLB) is 33% smaller than GLTF
embedded, which is handy for the asset download time. Also, GLB doesn't
require base64 decoding of the binary stuff, which is a plus. Also,
Blender makes you jump throgh hoops to actually get it to agree to
export GLTF embedded, but always defaults to GLB, so a few clicks saved
there too.

Internally, GLB is the same JSON prefixed by a binary header and offsets
and sizes for the JSON part and the binary part, which is buffer 0.

Implement support for loading GLBs as GLTFs.

Signed-off-by: Alexander Shishkin <alex@ash.works>