util: Make darray_delete() free the array when the last element is gone
Currently, darray_resize() only grows the array, but does nothing if
called from darray_delete(). Shrinking the array on each deletion might
be excessive, but it has to happen at some point. This may depend on the
element size or add/delete ratio, to be clever about it.

But at the bare minimum, darray_resize() should free the array when the
requested size is zero, which is what happens when you delete the last
element. Do that.

Signed-off-by: Alexander Shishkin <alex@ash.works>