character: Remove manually setting character's model matrix
The original idea seems to have been to derive rotation from the physics
simulation while the body is airborne, which then got replaced with an
open-coded version of what default_update() does, and will do anyway at
the bottom of the function via c->orig_update(...). Remove the
open-coded version.

Signed-off-by: Alexander Shishkin <alex@ash.works>