util: Unify parameters of darray_*() calls
Some are functions and take a pointer to struct darray, some are macros
and take a pointer to the anonymous union, struct darray or just the
union. This makes the API annoying, because one can never remember which
is which.

Turn all of them into macros that take the anonymous union (not a
pointer). While at it, make darray_get(), darray_add(), darray_insert()
return the pointer to the array's element type.

Having done that, convert all users.

Signed-off-by: Alexander Shishkin <alex@ash.works>