character, scene: Move motion controller to the scene
Motion controller that converts inputs into character and camera motion
direction doesn't belong to each character, but instead to the scene.
It can control only one character (although there may be multiple of
them (2/split screen, maybe) in the future). Move the motion controller
to the scene and let the character that's currently controlled use it.

Signed-off-by: Alexander Shishkin <alex@ash.works>