model: Use entity3d position update helpers
There are multiple places in the code where an entity's position gets
changed/updated, and there are helpers for that already. Change the
users to use the helpers instead. At the same time, add
entity3d::updated field to be used in needs_update() instead of
comparing old position updates against the cached ones, use that, and
set it in the position update helpers.

Also, phys_body::updated is added to avoid phys_body position update
when entity3d_position()/entity3d_move() are called with phys_body's
position.

Signed-off-by: Alexander Shishkin <alex@ash.works>