clap: Add a wrapper for the frame callback function
A lot of code in the render_frame() function belongs in the engine core,
so start a clap_frame() and move fps calculation code there and fps
calculation data out of struct scene into clap_context. Also, add
clap_resize(), because they share callback data (because the display.h
interface is like that).

Signed-off-by: Alexander Shishkin <alex@ash.works>