character: Fix double jumping
When a character jumps and lands it immediately jumps again, unless the
jump button press is frame perfect, because the jump setting that comes
from the input handler doesn't get cleared while the character is
airborne. But it only needs to be observed between the input handler and
character_move() to be effective.

Commit e52165b ("character: Move dashing and jumping out of struct
motionctl") removed clearing the jump setting from motion_reset(), where
it worked correctly every frame, but forgot to add it to
character_motion_reset(), thus introducing the double jump. Fix this by
clearing character::jump in character_motion_reset().

Signed-off-by: Alexander Shishkin <alex@ash.works>
Fixes: #74