font: Fix font name assignment
Commit b295cd5 ("font: Load fonts via lib_read_file()") removed the
local variable font_name that was used in asprintf()ing the font->name
on load. The compilation didn't break because there's also a function
called font_name(), but since it's not a NUL-terminated string, the
mem_asprintf() crashes trying to take a strlen() of it. Use
font_open()'s parameter "name" instead.

Signed-off-by: Alexander Shishkin <alex@ash.works>