ui-animations: Rework sequencing
To make it clear if a UI animation should wait for the previous one to
complete or run in parallel, add "wait" parameter to the animation
functions that run for multiple frames. If it's true, let the previous
animation play out before this one starts, otherwise run them in
parallel. Update the API and all the users to reflect this.

While at it, make ui-animations.c slightly less messy and add comments
explaining what's going on.

Signed-off-by: Alexander Shishkin <alex@ash.works>