ui: Fix transparency for text strings
The framebuffer texture where ui_render_string() renders glyphs is
plugged into a model that's using the "ui" shader, which uses a
different logic for color passthrough (overrides all texture's texel
colors with the in_color passed into the shader), which with
COLOR_PT_ALPHA produces a text on a black background when alpha is
greater than zero, because all the textels that don't belong to a glyph
are black (but their alpha component is 0, so in a normal situation they
are transparent).

Switch this model to use the "glyph" shader that only changes the alpha
component on non-black pixels, so that only the glyphs can get non-zero
alpha, allowing for text fade-in or fade-out or any other use of alpha
blending without showing the black background.

Signed-off-by: Alexander Shishkin <alex@ash.works>