ui-animations: Use time instead of frames
Framerate is prone to changing, which slows down animation execution,
whereas time flows at constant rate, which means that regardless of
framerate drops, the animations will take the same time to play out.
Change the animations' notion of time to the actual time. This also
changes their APIs, so update the users. This leaves the problem of
delivering the current time to the ui object in a somewhat hacky state,
which will be solved in the course of finalizing issue #47.

One unfortunate side effect of this is that it becomes impossible to
debug these animations frame-by-frame using breakpoints, but it's
already quite cumbersome anyway.

Signed-off-by: Alexander Shishkin <alex@ash.works>
Fixes: #84