scene: Don't leak model shader reference
In scene_camera_add(), a shader is looked up by shader_prog_find(),
which bumps the shader's reference count and this shader is then passed
to model3d_new...(), which takes another reference, but the first one is
never released. Pass the reference of the found shader into model3d_new,
so it can take the ownership of it using ref_pass(), which exists
specifically for this purpose.

Signed-off-by: Alexander Shishkin <alex@ash.works>