windows-libc: Add stubs for potentially missing Rtl* functions
The documentation for RtlUnicodeStringToUTF8String() says "minimum
supported client" is "Windows 10, version 2004", but github's
windows-2019 is older than that. windows-libc uses these functions left
right and center, which causes executables linked with it to crash with
0xc0000139 exit code, which was a joy to debug.

Add stub implementations of these functions as static inlines in
internal/nt.h, so that resulting executables can still run. They only
work if the strings involved are ASCII, but that's good for now.

While at it, fix stdio cleanup to actually remove its streams.

Signed-off-by: Alexander Shishkin <alex@ash.works>
Fixes: #79