render/gl: Add binding_points_t
uniform_buffer_t needs to be bound to binding points in the shader(s).
This process is differs depending on a renderer. Add an abstraction to
allow higher level code to do it in a renderer-agnostic way. Add a GL
implementation, where uniforms/uniform buffers are bound to the entire
program.

Signed-off-by: Alexander Shishkin <alex@ash.works>