render/gl: Probe for supported texture formats
The only way to know for sure which formats a driver supports (like
WebGL) is to probe for them. Test each supported format for standalone
textures and framebuffer attechment textures. Add APIs to query if a
format is supported for either purpose from the result of this probing.
Validate requested formats in texture_init(), texture_load() and
fbo_new().

Signed-off-by: Alexander Shishkin <alex@ash.works>