model: Clean up models_render()
models_render() has a monstrous number of arguments, most of which are
optional and quite a few are not in use, some of which for a long time.
Clean up the argument list, remove unused uniform settings etc. Derive
projection matrix from either camera or light if it's needed.

Update all callers.

Signed-off-by: Alexander Shishkin <alex@ash.works>