model: Make most of the models_render() arguments optional
At the minimum, models_render() requires a renderer object and a model
queue, everything else (light, camera, shader_override, cascade,
entity_count, txm_count) is optional, with the exception that if you
have either light or camera, you also need to specify the cascade. Move
all these into an options structure and pass it in via a variadic macro.

This also allows for adding new options without having to do a treewide
sweep to add another NULL to 95% of the callers.

Update callers.

Signed-off-by: Alexander Shishkin <alex@ash.works>