cmake: Make builddate.c regenerate on every build
Currently, CONFIG_BUILDDATE and builddate.c contain configuration time,
not build time, which just tells you that you may have touched a cmake
file in the last few days or something. So, to remedy this, force
generate builddate.c on EVERY build however incremental. This puts
CONFIG_BUILDDATE out of sync with build_date string, so just remove it,
there's no reason it has to be a macro, it'll end up a string constant
either way. This way it'll happen only once, though.

Signed-off-by: Alexander Shishkin <alex@ash.works>