render, pipeline: Add support for MRT to textures
Currently, the render-gl.c/render.h code works under the assumption that
multiple render targets (MRT) can only be had with render buffer
attachments and that there can only be one texture attachment, either
depth or color. This is not the case, multiple texture attachments are
totally possible within one framebuffer.

First, rework the fbo_attachment type to reflect this possibility, since
it's used to both specify individual attachments and the entire
framebuffer configuration. Change it from an enumeration type to a
bitmask wrapped in a union (partly for the reason that accidentally left
over arithemic operations on it would be rejected by the compiler,
leaving fewer obscure edge case bugs to deal with later).

To accomodate this type, change the APIs, add necessary helper functions
and macros.

Update the pipeline code to work with this new type, because a
render_pass configuration equals framebuffer configuration.

Update the pipeline adjacent modules like pipeline-builder to switch to
the new type and its APIs and in particular switch the currently only
MRT pass (model pass) to textures instead of render buffers.

Finally, update the one FBO user outside of the aforementioned modules.

Signed-off-by: Alexander Shishkin <alex@ash.works>