shaders: Fix model's shadow factor calculation
It was originally assumed that MSAA shadows were broken, because they
appeared pitch black as opposed to sensibly grey shadows of the PCF. But
turns out, there was a logical bug in the PCF implementation that didn't
include the rightmost columnt and the bottom row in the total, so the
total was always (except for shadow edges) 16.0/25.0 (kernel 5), which
inadvertently resulted in nice grey shadows. Fix that.

Now that that's fixed, PCF shadows are also pitch black. Of the variety
of ways of dealing with that, use a shadow factor multiplier that is
derived from the angle between the surface and the light vector, in
anticipation that the light source will be moving and the intensity of
shadows should change accordingly.

Signed-off-by: Alexander Shishkin <alex@ash.works>