pipeline: Add an option to enable MSAA for model pass
Currently, MSAA is enabled by default for the model render pass, even
though not all platforms handle it well and some downright don't support
it (in which case render-gl.c will ignore it), but they still have to
blit from the model render pass textures to access them.

Add an opt-in option to enable it.

Signed-off-by: Alexander Shishkin <alex@ash.works>