shaders, pipeline: Add Laplace edge filter
As an alternative to Sobel edge filters that fall back on Laplace for
depth edges, add a pure Laplace filter for both normal and depth edges
that produces a uniform result between the two. It doesn't have any
advantage with MSAA textures' sample averaging or any other sampling
trickery, so the same shader is used in both MSAA and non-MSAA cases.

Add a render option to use sobel and a make Laplace the default. Add a
debug UI switch between the two.

Signed-off-by: Alexander Shishkin <alex@ash.works>