character, shaders: Encode character ID in the edge input's alpha channel
Characters are treated specially in the normal edge map: teir diffuse
textures are used instead of normals to outline their texture details
instead of individual faces. Their colors, however clash with normals of
faces and each other. A hack is employed to remedy that, that's based on
an entity's address hash, which is also not reliable.

Instead, encode each character's ID in the alpha component of the
normal map and use it in the edge filters to push characters' texture
colors outside of [0.0, 1.0] range and into their own ranges, so they
don't clash with random surroundings or each other.

Signed-off-by: Alexander Shishkin <alex@ash.works>
Fixes: #107