sound: Add SFX support
Add basic support for sound effects. Use a simple linked list for a
container type for now, as there are not expected to be many of them at
the moment, but can be transparently upgraded when the need arises.
Containers can be embedded in the scene for global effects and models
for model-specific effects. The effects themselves may built in names to
sound files and contain said sound files to be played at designated
times.

Signed-off-by: Alexander Shishkin <alex@ash.works>
Fixes: #112