pipeline, ldjam56: Load shaders on demand in pipeline
Currently, the shaders are explicitly loaded in the demo/ldjam56 code
regardless of whether or not they're used in the pipeline or model
loading. This also means that the core code explicitly depends on the
game code to load correct shaders.

Instead, keep the shader list in the pipeline object and have the
pipeline code load shaders when render passes are created.

As a side effect, pass pipeline and mq objects instead of scene to gltf
loader.

Signed-off-by: Alexander Shishkin <alex@ash.works>