ui: Make UI health bar a progress bar widget
All the little widgets in ui.c are due for either removal or conversion
to ui_widget. The progress bar has multiple uses, so it goes first.
Convert the progress bar to a widget, move health bar instantiation from
ui.c to game.c that's its exclusive user (which is also currently
usused, so the ubiquitous health bar will disappear for the time being).

Signed-off-by: Alexander Shishkin <alex@ash.works>