physics: Cast rays from capsule's circumference for ground contact
When the bottom of the capsule collides with the ground, character's
height is corrected by the capsule offset and ground contact is returned
from phys_body_ground_collide(), but then the ray is cast from the
center of the capsule and may not collide with the ground, so if contact
happens at the edge, the character will keep rising and falling. Correct
this somewhat by casting 4 more rays from the circumference of the
capsule to establish ground contact.

Signed-off-by: Alexander Shishkin <alex@ash.works>